Cleanup some compiler warnings
git-svn-id: svn://svn.pgadmin.org/trunk@6744 a7884b65-44f6-0310-8a51-81a127f17b15

+++ pgadmin/ctl/ctlSQLBox.cpp
@@ -216,7 +216,7 @@ bool ctlSQLBox::DoFind(const wxString &find, const wxString &replace, bool doRep
 
                     // Stop if we've got to the end. This is important for the $
                     // case where it'll just keep finding the end of the line!!
-                    if (cr.cpMin + replace.Length() == GetLength())
+                    if ((int)(cr.cpMin + replace.Length()) == GetLength())
                         return false;
                 }
                 else
